+++ src/components/HeaderBlock/index.js
@@ -1,12 +1,19 @@
 import s from  './style.module.css';
 
-const HeaderBlock = ({ title, descr }) => {
+const HeaderBlock = ({ title, descr, onClickButton }) => {
+    const handleClick = () => {
+        onClickButton && onClickButton("game");
+    }
+
     return (
         <header className={s.root}>
             <div className={s.forest}></div>
             <div className={s.container}>
                 <h1>{title}</h1>
                 <p>{descr}</p>
+                <button onClick={handleClick}>
+                    Start Game
+                </button>
             </div>
         </header>
     )
