+++ src/components/LayoutBlock/index.js
@@ -1,3 +1,6 @@
+
+import cn from 'classnames';
+
 import s from  './style.module.css';
 
 const LayoutBlock = ({ id, title, urlBg, colorBg, children } ) => {
@@ -13,7 +16,7 @@ const LayoutBlock = ({ id, title, urlBg, colorBg, children } ) => {
                         { title && (<h3> {title} </h3>) }
                         <span className= {s.separator}></span>
                     </div>
-                    <div className= { `${s.desc} ${s.full}` }>
+                    <div className= { cn(s.desc, s.full) }>
                         { children }
                     </div>
                 </article>
