+++ src/components/MenuHeader/Menu/index.js
@@ -1,0 +1,43 @@
+import cn from 'classnames';
+
+import s from  './style.module.css';
+
+const MenuBlock = ({isActive}) => {
+    // const handleClick = () => {
+    //     onClickButton && onClickButton("game");
+    // }
+
+    return (
+        <div className={cn(s.menuContainer, {[s.active] : isActive}, {[s.deactive] : !isActive})}>
+          <div className={s.overlay} />
+          <div className={s.menuItems}>
+            <ul>
+              <li>
+                <a href="#welcome">
+                  HOME
+                </a>
+              </li>
+              <li>
+                <a href="#game">
+                  GAME
+                </a>
+              </li>
+              <li>
+                <a href="#about">
+                  ABOUT
+                </a>
+              </li>
+              <li>
+                <a href="#contact">
+                  CONTACT
+                </a>
+              </li>
+            </ul>
+          </div>
+        </div>
+    )
+};
+
+export default MenuBlock;
+
+
