+++ src/components/PokemonCard/index.js
@@ -1,28 +1,31 @@
 import { useState } from 'react';
 
+import CardBackedPicture from "../../assets/card-back-side.jpg";
+
+import cn from 'classnames';
+
 import s from  './style.module.css';
 
-import CardBackedPicture from "../../assets/card-back-side.jpg";
 
 const PokemonCard = ({id, name, img, type, values}) => {  
     const [isActive, setActive] = useState(false);
 
     const onClick = () => {
         setActive(!isActive);
-    }
+    };
 
     if(id && name && img && type && values) {
     return ( 
         <div className={s.root} onClick={onClick}>
-            <div className={`${s.pokemonCard} ${isActive ? s.active : ""}`}>
+            <div className={cn(s.pokemonCard, {[s.active] : isActive})}>
                 <div className={s.cardFront}>
-                    <div className={`${s.wrap} ${s.front}`}>
+                    <div className={cn(s.wrap, s.front)}>
                         <div className={`${s.pokemon} ${s.[type]}`} >
                             <div className={s.values}>
-                                <div className={`${s.count} ${s.top}`}> {values.top} </div>
-                                <div className={`${s.count} ${s.right}`}> {values.right} </div>
-                                <div className={`${s.count} ${s.bottom}`}> {values.bottom} </div>
-                                <div className={`${s.count} ${s.left}`}> {values.left} </div>
+                                <div className={cn(s.count, s.top)}> {values.top} </div>
+                                <div className={cn(s.count, s.right)}> {values.right} </div>
+                                <div className={cn(s.count, s.bottom)}> {values.bottom} </div>
+                                <div className={cn(s.count, s.left)}> {values.left} </div>
                             </div>
                             <div className={s.imgContainer}>
                                 <img src={img} alt={name} />
@@ -41,7 +44,7 @@ const PokemonCard = ({id, name, img, type, values}) => {
                 </div>
 
                 <div className={s.cardBack}>
-                    <div className={`${s.wrap} ${s.back}`}>
+                    <div className={cn(s.wrap, s.back)}>
                         <img src={CardBackedPicture} alt="Сard Backed" />
                     </div>
                 </div>
